move determine_device to seperate model_utils file
