+++ pyrelational/models/ensemble_model.py
@@ -7,8 +7,9 @@ from pytorch_lightning import LightningModule
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from torch.utils.data import DataLoader
 
+from ._model_utils import _determine_device
 from .generic_model import GenericModel
-from .lightning_model import LightningModel, _determine_device
+from .lightning_model import LightningModel
 
 
 class GenericEnsembleModel(GenericModel, ABC):
