+++ pyrelational/models/mcdropout_model.py
@@ -7,8 +7,9 @@ from pytorch_lightning import LightningModule
 from torch.nn.modules import Module
 from torch.utils.data import DataLoader
 
+from ._model_utils import _determine_device
 from .generic_model import GenericModel
-from .lightning_model import LightningModel, _determine_device
+from .lightning_model import LightningModel
 
 logger = logging.getLogger()
